requirements.txt
Django==4.1.3
django-crispy-forms==1.14.0
django-environ==0.8.1
mozilla-django-oidc==3.0.0
annotated-types==0.6.0
argcomplete==3.2.2
attrs==23.2.0
black==24.2.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
datamodel-code-generator==0.25.4
dnspython==2.6.1
email_validator==2.1.1
flower==2.0.1
genson==1.2.2
idna==3.6
inflect==5.6.2
isort==5.13.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
loguru==0.7.2
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
pydantic==2.6.2
pydantic_core==2.16.3
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.4
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
six==1.16.0
typing_extensions==4.10.0
urllib3==2.2.1
daphne==4.0.0
channels==4.0.0
channels-redis==4.1.0